+++ autopcr/http_server/httpserver_test.py
@@ -1,9 +1,6 @@
 from .httpserver import HttpServer
-from ..module.modules import register_all
 from ..module.crons import queue_crons
 import asyncio
-
-register_all()
 
 server = HttpServer(port=13200)
