+++ autopcr/module/crons.py
@@ -1,9 +1,6 @@
 import asyncio
 import datetime
-import os
-from ..module.modulebase import ModuleManager
 from ..module.accountmgr import instance as accountmgr
-from ..constants import CONFIG_PATH
 
 async def _cron(task):
     hour = datetime.datetime.now().hour
