+++ autopcr/module/modules/room.py
@@ -1,0 +1,108 @@
+from ..modulebase import *
+from ..config import *
+from ...core.pcrclient import pcrclient
+from ...model.error import *
+from ...db.database import db
+from ...model.enums import *
+from typing import List
+from ...model.requests import SendGiftData
+
+@description('收取家园体')
+@default(True)
+class room_accept_all(Module):
+    async def do_task(self, client: pcrclient):
+        room = await client.room_start()
+        for x in room.user_room_item_list:
+            if x.item_count:
+                res = await client.room_accept_all()
+                msg = await client.serlize_reward(res.reward_list)
+                self._log(msg)
+                return
+        raise SkipError('没有可收取的家园物品。')
+
+@description('升级家园家具')
+@default(True)
+class room_upper_all(Module):
+    async def do_task(self, client: pcrclient):
+        room = await client.room_start()
+        floors = {}
+        cnt = 0
+        for layout in room.room_layout.floor_layout:
+            cnt += 1
+            for item in layout.floor:
+                floors[item.serial_id] = cnt
+
+        for x in room.user_room_item_list:
+            if db.is_room_item_level_upable(client.data.team_level, x):
+                await client.room_level_up_item(floors[x.serial_id], x)
+                self._log(f"开始升级{db.get_inventory_name_san((eInventoryType.RoomItem, x.room_item_id))}至{x.room_item_level + 1}级")
+
+        if not self.log:
+            raise SkipError('没有可升级的家园物品。')
+
+@description('公会小屋点赞，先回赞，再随机点赞')
+@default(True)
+class room_like_back(Module):
+    async def do_task(self, client: pcrclient):
+        await client.room_start()
+        result = []
+        like_user = []
+        like_history = await client.room_like_history()
+        cnt = like_history.today_like_count
+        pos = 0
+        if cnt >= 10:
+            raise SkipError(f"今日已点赞{cnt}次")
+        while pos < like_history.today_be_liked_count or cnt < 10:
+            viewer_id = 0
+            if pos < like_history.today_be_liked_count:
+                viewer_id = like_history.like_history[pos].viewer_id
+            user = await client.room_visit(viewer_id)
+            pos += 1
+            if user.room_user_info.today_like_flag:
+                continue
+            try:
+                resp = await client.room_like(user.room_user_info.viewer_id)
+            except Exception as e:
+                if str(e).startswith("此玩家未读取的点赞数"):
+                    continue
+                else:
+                    raise(e)
+            result += resp.reward_list
+            like_user.append(user.room_user_info.name)
+            cnt += 1
+
+        result = await client.serlize_reward(result)
+        self._log(f"为【{'|'.join(like_user)}】点赞，获得了:\n" + result)
+
+@description('喂蛋糕')
+@default(True)
+class love_up(Module):
+    async def do_task(self, client: pcrclient):
+        for unit in client.data.unit_data.values():
+            unit_id = unit.chara_id * 100 + 1
+            unit_name = db.get_inventory_name_san((eInventoryType.Unit, unit_id))
+            love_level, total_love = db.max_total_love(client.data.unit[unit_id].unit_rarity)
+            if unit.chara_love < total_love:
+                dis = total_love - unit.chara_love
+                cakes: List[SendGiftData] = []
+                for cake in db.love_cake:
+                    current_num = client.data.get_inventory((eInventoryType.Item, cake.item_id))
+                    if not current_num:
+                        continue
+                    item_num = min(current_num, (dis + cake.value - 1) // cake.value)
+                    dis -= item_num * cake.value
+                    use_cake = SendGiftData()
+                    use_cake.item_id = cake.item_id
+                    use_cake.item_num = item_num
+                    use_cake.current_item_num = current_num
+                    cakes.append(use_cake)
+                    if dis <= 0: 
+                        break
+                if dis > 0:
+                    self._log(f"{unit_name}: 蛋糕数量不足")
+                    break
+                await client.multi_give_gift(unit_id, cakes)
+                self._log(f"{unit_name}: 亲密度升至{love_level}级")
+
+        if not self.log:
+            raise SkipError("所有角色均已亲密度满级")
